Move crosshatch function  back into operator file
